Version 0.5.2: added memofile validity tests
PgDBF now checks that memofiles are at least minimally correct before
attempting to use them. As of now, that means that it looks at the "nextblock"
field to see if it's negative or if it would extend far past the end of the
file. In either case, either the memofile is damaged or it isn't really
a memofile at all.