Version 0.5.4: Better fix for invalid memofiles
This version removes the somewhat ad-hoc method of checking a memofile's
length at the beginning of the run, and adds a check for valid memofile
offsets each time a memo record is requested. The overhead is negligible
and it adds the extra benefit of allowing as much data as possible to be
extracted from damaged memofiles.